engines/physics_engines/drake/python/requirements.txt
# Core scientific computing libraries
numpy
pandas==2.2.3
matplotlib==3.9.1
scipy

# Testing and development
pytest==8.2.1
pytest-cov==5.0.0
pytest-mock==3.14.0

# Code quality and formatting
ruff==0.5.0
black==24.4.2
mypy==1.19.0
pre-commit==4.5.0

# Configuration and utilities
pyyaml==6.0.1
pathlib2==2.3.7

# Type stubs for better type checking
types-PyYAML==6.0.12.12

# ============================================================================
# Modeling Stack (Pinocchio, Pink, Drake, QP solvers)
# ============================================================================
# Pinocchio - Analytical dynamics library
# Using compatible release constraint for stability while allowing patch updates
pin>=2.6.0

# Pink (pin-pink) - Stacked on top of Pinocchio for advanced kinematics
# Compatible with Pinocchio 2.6.x
pin-pink>=1.0.0

# Drake - Robot manipulation and control
# Drake releases frequently, using compatible release for latest 1.x
drake>=1.22.0

# QP Solvers - Quadratic programming solvers
qpsolvers
osqp>=0.6.3
scs~=3.2.4
quadprog~=0.1.11
PyQt6>=6.4.0,<7.0.0
